# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FitTrack API — fitness tracking, nutrition logging, and social features
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workouts:
    get:
      operationId: listWorkouts
      tags: [workouts]
      summary: List my workouts
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Workout list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workout"
    post:
      operationId: createWorkout
      tags: [workouts]
      summary: Start a new workout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutInput"
      responses:
        "201":
          description: Created workout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
  /workouts/{workoutId}:
    get:
      operationId: getWorkout
      tags: [workouts]
      summary: Get workout detail with all exercises and sets
      parameters:
        - name: workoutId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workout detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutDetail"
    patch:
      operationId: updateWorkout
      tags: [workouts]
      summary: Update workout (finish it, add notes, etc.)
      parameters:
        - name: workoutId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutUpdate"
      responses:
        "200":
          description: Updated workout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
    delete:
      operationId: deleteWorkout
      tags: [workouts]
      summary: Delete workout
      parameters:
        - name: workoutId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /workouts/{workoutId}/exercises:
    post:
      operationId: addWorkoutExercise
      tags: [workouts]
      summary: Add an exercise to a workout
      parameters:
        - name: workoutId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutExerciseInput"
      responses:
        "201":
          description: Added exercise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutExercise"
  /workouts/{workoutId}/exercises/{workoutExerciseId}:
    delete:
      operationId: removeWorkoutExercise
      tags: [workouts]
      summary: Remove an exercise from a workout
      parameters:
        - name: workoutId
          in: path
          required: true
          schema:
            type: integer
        - name: workoutExerciseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /workouts/{workoutId}/exercises/{workoutExerciseId}/sets:
    post:
      operationId: addSet
      tags: [workouts]
      summary: Add a set to a workout exercise
      parameters:
        - name: workoutId
          in: path
          required: true
          schema:
            type: integer
        - name: workoutExerciseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetInput"
      responses:
        "201":
          description: Added set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutSet"
  /workouts/{workoutId}/exercises/{workoutExerciseId}/sets/{setId}:
    patch:
      operationId: updateSet
      tags: [workouts]
      summary: Update a set (weight, reps, completed)
      parameters:
        - name: workoutId
          in: path
          required: true
          schema:
            type: integer
        - name: workoutExerciseId
          in: path
          required: true
          schema:
            type: integer
        - name: setId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetUpdate"
      responses:
        "200":
          description: Updated set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutSet"
    delete:
      operationId: deleteSet
      tags: [workouts]
      summary: Delete a set
      parameters:
        - name: workoutId
          in: path
          required: true
          schema:
            type: integer
        - name: workoutExerciseId
          in: path
          required: true
          schema:
            type: integer
        - name: setId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Workout:
      type: object
      required: [id, name, startedAt, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        notes:
          type: ["string", "null"]
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: ["string", "null"]
        durationMinutes:
          type: ["integer", "null"]
        totalVolume:
          type: ["number", "null"]
        totalSets:
          type: ["integer", "null"]
        exerciseCount:
          type: ["integer", "null"]
        isFinished:
          type: boolean
        createdAt:
          type: string
          format: date-time
    WorkoutDetail:
      type: object
      required: [id, name, startedAt, exercises, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        notes:
          type: ["string", "null"]
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: ["string", "null"]
        durationMinutes:
          type: ["integer", "null"]
        totalVolume:
          type: ["number", "null"]
        totalSets:
          type: ["integer", "null"]
        isFinished:
          type: boolean
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutExercise"
        createdAt:
          type: string
          format: date-time
    WorkoutInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        routineId:
          type: integer
        notes:
          type: string
    WorkoutUpdate:
      type: object
      properties:
        name:
          type: string
        notes:
          type: string
        finishedAt:
          type: string
          format: date-time
        isFinished:
          type: boolean
    WorkoutExercise:
      type: object
      required: [id, exerciseId, exerciseName, order, sets]
      properties:
        id:
          type: integer
        exerciseId:
          type: integer
        exerciseName:
          type: string
        order:
          type: integer
        notes:
          type: ["string", "null"]
        sets:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutSet"
    WorkoutExerciseInput:
      type: object
      required: [exerciseId]
      properties:
        exerciseId:
          type: integer
        order:
          type: integer
        notes:
          type: string
    WorkoutSet:
      type: object
      required: [id, setNumber, completed]
      properties:
        id:
          type: integer
        setNumber:
          type: integer
        weight:
          type: ["number", "null"]
        reps:
          type: ["integer", "null"]
        completed:
          type: boolean
        setType:
          type: string
          enum: [normal, warmup, dropset, failure]
        previousWeight:
          type: ["number", "null"]
        previousReps:
          type: ["integer", "null"]
    SetInput:
      type: object
      required: [setNumber]
      properties:
        setNumber:
          type: integer
        weight:
          type: number
        reps:
          type: integer
        completed:
          type: boolean
        setType:
          type: string
          enum: [normal, warmup, dropset, failure]
    SetUpdate:
      type: object
      properties:
        weight:
          type: number
        reps:
          type: integer
        completed:
          type: boolean
        setType:
          type: string
          enum: [normal, warmup, dropset, failure]
...